+++ core/src/com/mygdx/clashofclans/Tokens/Warriors/Characters/Ringo.java
@@ -13,7 +13,7 @@ public class Ringo extends Beast {
     private static String PATH_TO_ATTACK_ANIMATION_L = "Ringo/Ringo L(Attack).gif";
     private static String PATH_TO_HURT_ANIMATION_L = "Ringo/Ringo L(Hurt).gif";
 
-    public Ringo(int pInitialX, int pInitialY) {
+    public Ringo(float pInitialX, float pInitialY) {
         super(pInitialX, pInitialY, new String[]{PATH_TO_IDLE_ANIMATION,
                 PATH_TO_WALKING_ANIMATION,
                 PATH_TO_ATTACK_ANIMATION,
